Chris/mock

See merge request auckland-ict-grad-school/final-project-groups-2020-s1/group-6/pgcit-project-fulltime-starter!5

+++ src/ictgradschool/project/entity/Article.java
@@ -4,5 +4,5 @@ public class Article {
     public int id;
     public String title;
     public String content;
-    public int date;
+    public long date;
 }
